Move Model3 project to correct folder, Adjusted namespaces (#453)
