+++ NetDaemon.sln
@@ -36,7 +36,9 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Integration", "Integration"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TestClient", "tests\Integration\TestClient\TestClient.csproj", "{CEF690C5-DF76-465D-B571-FBBC85C7F7BB}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Model3", "Model3\Model3.csproj", "{3C2E37A3-4669-4CF5-9D6C-D68FD2034642}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NetDaemon.Model3", "src\Model3\NetDeamon.Model3\NetDaemon.Model3.csproj", "{3C2E37A3-4669-4CF5-9D6C-D68FD2034642}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Model3", "Model3", "{1D0125A6-DDC5-43C1-AAD7-7BFE14AD863F}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -170,7 +172,7 @@ Global
 		{E33F4CFC-7D80-48C4-B5FE-107E36CD98E4} = {CD3259F6-4CD3-49AA-90D3-3CFEEE0BFF0E}
 		{374E37D9-68D9-4A65-ACBC-C6DC2A7C3076} = {32E20E48-2F62-447B-83A2-DB2EA7F61AB3}
 		{CEF690C5-DF76-465D-B571-FBBC85C7F7BB} = {374E37D9-68D9-4A65-ACBC-C6DC2A7C3076}
-		{3C2E37A3-4669-4CF5-9D6C-D68FD2034642} = {53E4E82B-565F-4D91-8C87-52B3C3B74884}
+		{3C2E37A3-4669-4CF5-9D6C-D68FD2034642} = {1D0125A6-DDC5-43C1-AAD7-7BFE14AD863F}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {7C5FBB7F-654C-4CAC-964F-6D71AF3D62F8}
