+++ src/DevelopmentApps/NetDaemon.DevelopmentApps.csproj
@@ -26,8 +26,9 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\..\Model3\Model3.csproj" />
     <ProjectReference Include="..\App\NetDaemon.App\NetDaemon.App.csproj" />
+    <ProjectReference Include="..\Daemon\NetDaemon.Daemon\NetDaemon.Daemon.csproj" />
+    <ProjectReference Include="..\Model3\NetDeamon.Model3\NetDaemon.Model3.csproj" />
   </ItemGroup>
 
   <ItemGroup>
