+++ src/DevelopmentApps/apps/CallServiceApp.cs
@@ -1,7 +1,8 @@
 ﻿using System.Text.Json.Serialization;
 using NetDaemon.Common;
-using NetDaemon.Common.ModelV3;
-using NetDaemon.Common.ModelV3.Domains;
+using NetDaemon.Model3.Common;
+using NetDaemon.Model3.Domains.Climate;
+
 
 namespace NetDaemon.DevelopmentApps.apps
 {
