+++ src/Model3/NetDeamon.Model3/Common/Context.cs
@@ -1,4 +1,4 @@
-﻿namespace NetDaemon.Common.ModelV3
+﻿namespace NetDaemon.Model3.Common
 {
     /// <summary>
     ///     Context
