+++ src/Model3/NetDeamon.Model3/DependencyInjectionSetup.cs
@@ -1,10 +1,10 @@
 ﻿using System;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
-using Model3.ModelV3;
-using NetDaemon.Common.ModelV3;
+using NetDaemon.Model3.Common;
+using NetDaemon.Model3.Internal;
 
-namespace Model3
+namespace NetDaemon.Model3
 {
     public static class DependencyInjectionSetup
     {
