+++ src/Model3/NetDeamon.Model3/Entities/StateChange.cs
@@ -1,6 +1,4 @@
-﻿using System;
-
-namespace NetDaemon.Common.ModelV3
+﻿namespace NetDaemon.Model3.Entities
 {
     public record StateChange()
     {
