+++ src/Model3/NetDeamon.Model3/NetDaemon.Model3.csproj
@@ -7,14 +7,16 @@
     </PropertyGroup>
 
     <ItemGroup>
-      <ProjectReference Include="..\src\Daemon\NetDaemon.Daemon\NetDaemon.Daemon.csproj" />
+      <PackageReference Include="JoySoftware.HassClient" Version="21.33.0-beta" />
+      <PackageReference Include="Microsoft.Extensions.Hosting.Abstractions" Version="5.0.0" />
+      <PackageReference Include="System.Reactive" Version="5.0.0" />
     </ItemGroup>
 
   <PropertyGroup>
-    <CodeAnalysisRuleSet>..\.linting\roslynator.ruleset</CodeAnalysisRuleSet>
+    <CodeAnalysisRuleSet>..\..\..\.linting\roslynator.ruleset</CodeAnalysisRuleSet>
   </PropertyGroup>
   <PropertyGroup>
-    <CodeAnalysisRuleSet>..\.linting\roslynator.ruleset</CodeAnalysisRuleSet>
+    <CodeAnalysisRuleSet>..\..\..\.linting\roslynator.ruleset</CodeAnalysisRuleSet>
     <EnforceCodeStyleInBuild>true</EnforceCodeStyleInBuild>
     <AnalysisMode>AllEnabledByDefault</AnalysisMode>
   </PropertyGroup>
